+++ src/easysam/validate_cloud.py
@@ -1,0 +1,5 @@
+from easysam.utils import get_aws_client
+
+
+def validate(cliparams: dict, resources_data: dict, errors: list[str]):
+    print('Not implemented')