+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 1.3.0
 
-(TBD)
+- Added `inspect schema` command to validate the resources.yaml file
+- Added `inspect resources` command to inspect required cloud resources
+- Added `inspect common-deps` command to inspect the common dependencies
 
 # 1.2.0